week10/requirements.txt
pip
Flask
cassandra-driver   --install-option="--no-cyton"
